extract AbstractXliffVersion
