+++ src/Symfony/Component/Translation/Loader/XliffVersion/XliffVersion12.php
@@ -11,7 +11,6 @@
 
 namespace Symfony\Component\Translation\Loader\XliffVersion;
 
-use DOMDocument;
 use Symfony\Component\Translation\MessageCatalogue;
 
 /**
@@ -19,7 +18,7 @@ use Symfony\Component\Translation\MessageCatalogue;
  *
  * @author Berny Cantos <be@rny.cc>
  */
-class XliffVersion12
+class XliffVersion12 extends AbstractXliffVersion
 {
     /**
      * Get validation schema source for this version
@@ -36,11 +35,11 @@ class XliffVersion12
     /**
      * Extract messages and metadata from DOMDocument into a MessageCatalogue
      *
-     * @param DOMDocument      $dom       Source to extract messages and metadata
+     * @param \DOMDocument     $dom       Source to extract messages and metadata
      * @param MessageCatalogue $catalogue Catalogue where we'll collect messages and metadata
      * @param string           $domain    The domain
      */
-    public function extract(DOMDocument $dom, MessageCatalogue $catalogue, $domain)
+    public function extract(\DOMDocument $dom, MessageCatalogue $catalogue, $domain)
     {
         $xml = simplexml_import_dom($dom);
         $encoding = strtoupper($dom->encoding);
@@ -80,58 +79,5 @@ class XliffVersion12
                 $catalogue->setMetadata((string) $source, array('notes' => $notes), $domain);
             }
         }
-    }
-
-    /**
-     * Internally changes the URI of a dependent xsd to be loaded locally
-     *
-     * @param string $schemaSource Current content of schema file
-     * @param string $xmlUri       External URI of XML to convert to local
-     *
-     * @return string
-     */
-    protected function fixXmlLocation($schemaSource, $xmlUri)
-    {
-        $newPath = str_replace('\\', '/', __DIR__).'/../schema/dic/xliff-core/xml.xsd';
-        $parts = explode('/', $newPath);
-        if (0 === stripos($newPath, 'phar://')) {
-            $tmpfile = tempnam(sys_get_temp_dir(), 'sf2');
-            if ($tmpfile) {
-                copy($newPath, $tmpfile);
-                $parts = explode('/', str_replace('\\', '/', $tmpfile));
-            }
-        }
-        $drive = '\\' === DIRECTORY_SEPARATOR ? array_shift($parts).'/' : '';
-        $newPath = 'file:///'.$drive.implode('/', array_map('rawurlencode', $parts));
-
-        return str_replace($xmlUri, $newPath, $schemaSource);
-    }
-
-    /**
-     * Convert a UTF8 string to the specified encoding
-     *
-     * @param string $content String to decode
-     * @param string $encoding Target encoding
-     *
-     * @throws \RuntimeException
-     * @return string
-     */
-    protected function utf8ToCharset($content, $encoding = null)
-    {
-        if (empty($encoding) || 'UTF-8' === $encoding) {
-            return $content;
-        }
-
-        if (function_exists('mb_convert_encoding')) {
-            return mb_convert_encoding($content, $encoding, 'UTF-8');
-        }
-
-        if (function_exists('iconv')) {
-            return iconv('UTF-8', $encoding, $content);
-        }
-
-        throw new \RuntimeException(
-            'No suitable convert encoding function (use UTF-8 as your encoding or install the iconv or mbstring extension).'
-        );
     }
 }
